Add verification step for object location in PRA scenario
