+++ tests/ctst/features/pra.feature
@@ -13,6 +13,7 @@ Feature: PRA operations
     And object "obj-2" should be "transitioned" and have the storage class "e2e-cold"
     And dmf volume should contain <objectCount> objects
     Given a DR installed
+    Then object "obj-1" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
 
     Examples:
     | versioningConfiguration | objectCount | objectSize |
